Refactor: invitedDashboardListType 분리
